+++ spi/contract-spi/src/main/java/org/eclipse/dataspaceconnector/spi/contract/negotiation/store/ContractNegotiationStore.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2021 Microsoft Corporation
+ *  Copyright (c) 2021 - 2022 Microsoft Corporation
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - add functionalities
  *
  */
 
@@ -19,6 +20,7 @@ import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
 import org.eclipse.dataspaceconnector.spi.system.Feature;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreement;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.negotiation.ContractNegotiation;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.stream.Stream;
@@ -71,8 +73,28 @@ public interface ContractNegotiationStore extends StateEntityStore<ContractNegot
      * <p>
      *
      * @param querySpec The query spec, e.g. paging, filtering, etc.
-     * @return A potentially empty stream of {@link ContractNegotiation}, never null.
+     * @return a stream of ContractNegotiation, cannot be null.
      */
+    @NotNull
     Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec);
 
+    /**
+     * Finds all contract agreements that are based on a specific contract definition
+     *
+     * @param definitionId the contract definition id
+     * @return a stream of ContractAgreement, cannot be null.
+     */
+    @NotNull
+    Stream<ContractAgreement> getAgreementsForDefinitionId(String definitionId);
+
+
+    /**
+     * Finds all contract agreement that are covered by a specific {@link QuerySpec}.
+     * If no {@link QuerySpec#getSortField()} is specified, results are not explicitly sorted.
+     *
+     * @param querySpec The query spec, e.g. paging, filtering, etc.
+     * @return a stream of ContractAgreement, cannot be null.
+     */
+    @NotNull
+    Stream<ContractAgreement> queryAgreements(QuerySpec querySpec);
 }